requirements.txt
pyserial>=3
numpy
scipy
imageio
scikit-image
dask
xarray
napari
qtpy
zarr
pyqt5
PySide2
